포인트 조회api

+++ src/main/java/com/cdy/ecommerce/ecommerce/api/admin/product/ProductAdminController.java
@@ -28,6 +28,6 @@ public class ProductAdminController {
 
     private ProductAdminDTO.Response entityToDTO(Product product) {
         // 변환
-        return ProductAdminDTO.Response.builder().id(product.getId()).name(product.getName()).price(product.getPrice().intValue()).build();
+        return ProductAdminDTO.Response.builder().id(product.getId()).name(product.getName()).price(product.getPrice()).build();
     }
 }
