+++ src/main/java/com/cdy/ecommerce/ecommerce/api/admin/product/ProductAdminDTO.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 
 public class ProductAdminDTO {
 
@@ -17,7 +16,7 @@ public class ProductAdminDTO {
     public static class Request {
         private String userId; // 상품을 등록하는 사용자의 ID
         private String name; // 상품 이름
-        private BigDecimal price; // 상품 가격
+        private int price; // 상품 가격
         private int initialStock; // 초기 재고 수량
     }
