+++ src/main/java/com/cdy/ecommerce/ecommerce/domain/product/business/model/Product.java
@@ -3,7 +3,7 @@ package com.cdy.ecommerce.ecommerce.domain.product.business.model;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.math.BigDecimal;
+
 
 @Entity
 @Table(name = "ecommerce_product")
@@ -22,7 +22,7 @@ public class Product {
     @Column(name = "product_name")
     private String name;
     @Column(name = "product_price")
-    private BigDecimal price;
+    private int price;
 
     @OneToOne(mappedBy = "product", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     private ProductStock productStock;
@@ -32,7 +32,7 @@ public class Product {
     public void changeName(String name){
         this.name = name;
     }
-    public void changePrice(BigDecimal price) {
+    public void changePrice(int price) {
         this.price = price;
     }
